requirements.txt
asgiref==3.8.1
awsebcli==3.20.10
boto3==1.34.109
botocore==1.31.85
Brotli==1.1.0
cement==2.8.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.3
distlib==0.3.8
dj-database-url==2.1.0
Django==5.0.6
django-storages==1.14.3
djangorestframework==3.15.1
drf-yasg==1.21.7
filelock==3.14.0
gunicorn==22.0.0
h11==0.14.0
idna==3.7
inflection==0.5.1
jmespath==1.0.1
packaging==24.0
pathspec==0.10.1
platformdirs==4.2.2
portalocker==2.8.2
psycopg2==2.9.9
psycopg2-binary==2.9.9
pypiwin32==223
python-dateutil==2.9.0.post0
pytz==2024.1
pywin32==306
PyYAML==5.3.1
requests==2.32.1
s3transfer==0.10.1
semantic-version==2.8.5
setuptools==69.5.1
six==1.16.0
sqlparse==0.5.0
termcolor==1.1.0
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==1.26.18
uvicorn==0.29.0
virtualenv==20.26.2
wcwidth==0.1.9
wheel==0.43.0
whitenoise==6.6.0
